even more refactoring and upgrades
:+1:  :+1:
