+++ static/put.js
@@ -1,10 +1,14 @@
 import { str2seed } from "./str2seed.js";
 
-function send(event) {
+/**
+ * Valids form
+ * 
+ * @returns {boolean}
+ */
+function validForm() {
     let error = document.getElementById("error");
     let seed = document.getElementById("seed").value;
     seed = seed.toUpperCase();
-    console.log(str2seed(seed.substring(0, 4) + " " + seed.substring(4, 8)));
     if (str2seed(seed.substring(0, 4) + " " + seed.substring(4, 8)) == 0) {
         error.innerHTML = 'Invalid seed';
         return false;
@@ -13,4 +17,4 @@ function send(event) {
     return true;
 }
 
-document.getElementById("seedForm").onsubmit = send;
+document.getElementById("seedForm").onsubmit = validForm;
