+++ static/str2seed.js
@@ -1,30 +1,39 @@
-export function str2seed(seed){
-    if(seed.length != 9)
+/**
+ * Converts seed in string to number seed
+ * 
+ * @param {string} seed 
+ * @returns {number}
+ */
+export function str2seed(seed) {
+    if(seed.length != 9) {
         return 0
+    }
     //"xxxx xxxx"
-    if(seed[4] != ' '){
+    if(seed[4] != ' ') {
         return 0
     }
 
     let dict = []
-    for(let i=0;i<255;i++){
+    for(let i=0;i<255;i++) {
         dict[i] = 0xFF
     }
-    for(let i=0;i<32;i++){
+    for(let i=0;i<32;i++) {
         dict["ABCDEFGHJKLMNPQRSTWXYZ01234V6789".charCodeAt(i)]=i
     }
 
     let num_seed = []
-    for(let i=0;i<9;i++){
-        if(i == 4)
+    for(let i=0;i<9;i++) {
+        if(i == 4) {
             continue
+        }
         let j = i
-        if(i > 4){
+        if(i > 4) {
             j = i-1
         }
         num_seed[j] = dict[seed.charCodeAt(i)]
-        if(num_seed[j] == 0xFF)
+        if(num_seed[j] == 0xFF) {
             return 0
+        }
     }
 
     let v8 = 0;
